bulssuk_ai/requirements.txt
fastapi==0.100.0
uvicorn==0.23.2
torch==2.0.1
ultralytics==8.0.124
pillow==9.5.0
pandas==2.1.1